Moved download logic to utils
