+++ src/components/Preview/Modal.tsx
@@ -11,12 +11,7 @@ import Ogma, { Size } from "@linkurious/ogma";
 import embedFonts from "@linkurious/svg-font-embedder";
 import { ImageViewer } from "../ImageViewer";
 import { Footer } from "./Footer";
-import {
-  downloadBlob,
-  getOgmaBackgroundColor,
-  scaleGraph,
-  stringToSVGElement,
-} from "../../utils";
+import { scaleGraph, stringToSVGElement } from "../../utils";
 import {
   addCheckerboard,
   addClipShape,
@@ -24,8 +19,7 @@ import {
   embedImages,
 } from "../../utils/svg";
 import { fullSizeMargin } from "../../constants";
-import { jsPDF } from "jspdf";
-import "svg2pdf.js";
+import { handleDownload } from "../../utils/download";
 
 // TODO: add that, and through the webworker
 //import { optimize } from "svgo/dist/svgo.browser";
@@ -120,49 +114,7 @@ export const Modal: FC<Props> = ({ visible, onCancel, onOk }) => {
     const el = stringToSVGElement(image as string);
     const bg = el.querySelector(".ogma-svg-background") as SVGRectElement;
     bg!.setAttribute("fill-opacity", background ? "1" : "0");
-    if (format.label === "PDF") {
-      const svg = el.cloneNode(true) as SVGSVGElement;
-      const svgWidth = parseInt(svg.getAttribute("width") as string);
-      const svgHeight = parseInt(svg.getAttribute("height") as string);
-      const pdf = new jsPDF({
-        format: [svgWidth, svgHeight],
-        orientation: svgWidth > svgHeight ? "landscape" : "portrait",
-        unit: "pt",
-        compress: true,
-      });
-      const margin = 0;
-      const pageWidth = pdf.getPageWidth() - margin * 2;
-      const pageHeight = pdf.getPageHeight() - margin * 2;
-
-      // vertical cursor
-      let y = margin;
-      // fitting ratio canvas to page
-      let ratio = 1 / Math.max(svgWidth / pageWidth, svgHeight / pageHeight);
-      // seems like big sizes breaks it
-      const width = svgWidth * ratio;
-      const height = svgHeight * ratio;
-      // resize SVG
-      el.setAttribute("width", width.toString());
-      el.setAttribute("height", height.toString());
-      // fit the contents of SVG
-      svg.setAttribute("viewBox", `0 0 ${width / ratio} ${height / ratio}`);
-
-      await pdf.svg(svg, { x: margin, y, width, height });
-      await pdf.save(`${visualisation.title}.pdf`);
-    } else {
-      const imgDownload = svgElementToString(el);
-
-      if (format.label === "SVG") {
-        downloadBlob(
-          imgDownload,
-          `${visualisation.title}.svg`,
-          "image/svg+xml"
-        );
-      } else if (format.label === "PNG") {
-        const data = await svgToPng(el);
-        downloadBlob(data, `${visualisation.title}.png`, "image/png");
-      }
-    }
+    await handleDownload(el, format, visualisation.title);
     if (onOk) onOk();
   };
